added with confirmation box

+++ client/src/pages/booking/BookingList.jsx
@@ -5,6 +5,8 @@ import { MdDelete } from "react-icons/md";
 import { FaSort, FaSortUp, FaSortDown } from 'react-icons/fa';
 import bookingService from "../../services/bookingService";
 import { toast } from 'react-toastify';
+import { confirmAlert } from "react-confirm-alert";
+import "react-confirm-alert/src/react-confirm-alert.css"; // Importing css
 
 export const BookingList = ({ BookingResponse }) => {
 
@@ -52,18 +54,44 @@ export const BookingList = ({ BookingResponse }) => {
     setSortConfig({ key, direction });
   };
 
-  const deleteBooking = (bookingId) => {
-    bookingService.deleteBooking(bookingId)
-      .then(response => {
-        toast.success("Booking deleted successfully");
-        window.location.reload();
-      })
-      .catch(error => {
-        toast.error("Failed to delete booking");
-      });
-  }
+  // const deleteBooking = (bookingId) => {
+  //   bookingService.deleteBooking(bookingId)
+  //     .then(response => {
+  //       toast.success("Booking deleted successfully");
+  //       window.location.reload();
+  //     })
+  //     .catch(error => {
+  //       toast.error("Failed to delete booking");
+  //     });
+  // }
 
-  return (
+
+  const deleteBooking = (bookingId) => {
+    confirmAlert({
+      title: "Confirm Delete",
+      message: "Are you sure you want to delete this booking?",
+      buttons: [
+        {
+          label: "Yes",
+          onClick: () => {
+            bookingService.deleteBooking(bookingId)
+              .then(response => {
+                toast.success("Booking deleted successfully");
+                window.location.reload(); // Reloading the page to reflect changes
+              })
+              .catch(error => {
+                toast.error("Failed to delete booking");
+              });
+          }
+        },
+        {
+          label: "No",
+          onClick: () => {}
+        }
+      ]
+    });
+  }
+    return (
     <>
       <div>
         <Table striped bordered hover>
